pyproject.toml
[tool.poetry]
name = "projeto-final"
version = "0.1.0"
description = ""
authors = ["Fermoreno <fernando.santosm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<=3.12"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pyside6 = "^6.6.3.1"
numpy = "^1.26.4"
pytesseract = "^0.3.10"
matplotlib = "^3.8.4"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
